backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
google-genai
supabase==2.18.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.0.0